arreglo error de axel
